Remove the usage of the deprecated KotlinGradleSubplugin and use the new replacement instead.
Fixes #30
